chore: restructure project layout
- Ref: https://github.com/golang-standards/project-layout
- Allow one line installation with `go install`
- go mod tidy
