+++ .gitignore
@@ -1,2 +1,2 @@
-countryfetch-go
-countryfetch
+/countryfetch-go
+/countryfetch
