+++ README.md
@@ -11,15 +11,15 @@ It uses https://restcountries.com/ for back-end.
 Make sure you have [GO](https://go.dev/) installed and run this one-line installer:
 
 ```bash
-git clone https://github.com/CondensedMilk7/countryfetch-go.git && cd ./countryfetch-go && go install
+go install github.com/CondensedMilk7/countryfetch-go/cmd/countryfetch@latest
 ```
 
 If you have the original `countryfetch` and you want to keep it, do this instead:
 
 ```bash
-git clone https://github.com/CondensedMilk7/countryfetch-go.git && cd ./countryfetch-go && go build -o countryfetch-go && cp ./countryfetch-go ~/.local/bin/
-
+git clone https://github.com/CondensedMilk7/countryfetch-go.git && cd ./countryfetch-go && go build ./cmd/countryfetch/ -o countryfetch-go && cp ./countryfetch-go ~/.local/bin/
 ```
+
 First make sure you run `countryfetch -sync`, optionally with `-flag` to store data in cache.
 
 ```
