+++ go.mod
@@ -1,12 +1,14 @@
-module github.com/CondensedMilk7/countryfetch
+module github.com/CondensedMilk7/countryfetch-go
 
 go 1.19
+
+require github.com/qeesung/image2ascii v1.0.1
 
 require (
 	github.com/aybabtme/rgbterm v0.0.0-20170906152045-cc83f3b3ce59 // indirect
 	github.com/mattn/go-isatty v0.0.17 // indirect
 	github.com/nfnt/resize v0.0.0-20180221191011-83c6a9932646 // indirect
-	github.com/qeesung/image2ascii v1.0.1 // indirect
+	github.com/stretchr/testify v1.8.1 // indirect
 	github.com/wayneashleyberry/terminal-dimensions v1.1.0 // indirect
-	golang.org/x/sys v0.0.0-20220811171246-fbc7d0a398ab // indirect
+	golang.org/x/sys v0.4.0 // indirect
 )
